requirements.txt
langchain>=0.3.9
langchain-community>=0.3.9
langchain-huggingface>=0.1.2
langchain-google-genai>=2.0.6
python-dotenv>=1.0.1
streamlit>=1.38.0
tiktoken>=0.8.0
faiss-cpu>=1.9.0
protobuf>=5.29.0
nltk>=3.9.1
huggingface-hub>=0.26.3
InstructorEmbedding>=1.0.1
sentence-transformers>=3.3.1
streamlit>=1.40.2